chore: move input.ts to constants.ts in test util
