+++ src/pages/account/index.test.tsx
@@ -3,8 +3,11 @@ import { getAddress } from 'viem';
 import { useAccount } from 'wagmi';
 
 import AccountProfilePage, { Head } from '.';
-import { render, getMockSiteMetadata } from '../../utils/test-utils';
-import { VALID_ACCOUNT_1 } from '../../utils/test-utils/input';
+import {
+  render,
+  getMockSiteMetadata,
+  VALID_ACCOUNT_1,
+} from '../../utils/test-utils';
 
 jest.mock('../../features/account/components/AccountInfo', () => ({
   AccountInfo: () => <div />,
