+++ src/utils/test-utils/constants.ts
@@ -2,3 +2,5 @@ export const VALID_ACCOUNT_1 = '0x6B24aE0ABbeb67058D07b891aF415f288eA57Cc7';
 export const VALID_ACCOUNT_2 = '0x7B17936E2a44C1fFef392d13e490fc8426d3329e';
 
 export const INVALID_ADDRESS = '0x123123123';
+
+export const SNAP_SHASUM_1 = 'ALj0YMDz4JNchPJfreWErQx/GOh4FMWWy1o1rCq7sQQ=';
