+++ src/utils/test-utils/index.ts
@@ -3,3 +3,4 @@ export * from './queries';
 export * from './render';
 export * from './request';
 export * from './state';
+export * from './constants';
